Rename `useUpdateDemoSite`
